refactor: Component의 세부 package 만들어서 정리
