--- app/src/main/java/com/ganaljigi/kubf/ui/helper/component/ShortCutComponent.kt
@@ -1,2 +1,0 @@
-package com.ganaljigi.kubf.ui.helper.component
-
