+++ app/src/main/java/com/ganaljigi/kubf/ui/helper/component/information/InformationComponent.kt
@@ -1,0 +1,9 @@
+package com.ganaljigi.kubf.ui.helper.component.information
+
+//정보 제목 박스
+
+//정보 하나 박스
+
+//지도 박스
+
+//전체 정보 박스
