+++ app/src/main/java/com/ganaljigi/kubf/ui/helper/component/notice/NoticeComponent.kt
@@ -1,18 +1,18 @@
-package com.ganaljigi.kubf.ui.helper.component
+package com.ganaljigi.kubf.ui.helper.component.notice
 
-import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.KeyboardArrowLeft
+import androidx.compose.material.icons.filled.KeyboardArrowRight
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextFieldDefaults.contentPadding
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
@@ -20,17 +20,46 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
-//장애학생지원센터 TopAppBar
+//Helper화면의 NoticeBox
+@Composable
+fun NoticeBox(
+    onNavigateClick: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 16.dp)
+                .height(52.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = "공지사항",
+                fontWeight = FontWeight.SemiBold
+            )
+
+            IconButton(onClick = onNavigateClick) {
+                Icon(
+                    imageVector = Icons.Default.KeyboardArrowRight,
+                    contentDescription = "공지사항 페이지로 이동"
+                )
+            }
+        }
+    }
+}
+
+//Notice화면의 공지사항 TopAppBar
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun HelperTopAppBar(
+fun NoticeTopAppBar(
     onBackClick: () -> Unit
 ) {
     TopAppBar(
@@ -39,7 +68,7 @@ fun HelperTopAppBar(
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 Text(
-                    text = "장애학생지원센터",
+                    text = "공지사항",
                     fontSize = 20.sp,
                     fontWeight = FontWeight.SemiBold,
                     color = Color.Black,
@@ -61,14 +90,14 @@ fun HelperTopAppBar(
         ),
         modifier = Modifier
             .fillMaxWidth()
-            .height(64.dp)
+            .height(48.dp)
     )
 }
 
 @Preview(showBackground = true)
 @Composable
-fun PreviewHelperTopAppBar() {
-    HelperTopAppBar(
+fun PreviewNoticeTopAppBar() {
+    NoticeTopAppBar (
         onBackClick = {}
     )
 }
