+++ app/src/main/java/com/ganaljigi/kubf/ui/helper/component/shortcut/ShortCutComponent.kt
@@ -1,0 +1,7 @@
+package com.ganaljigi.kubf.ui.helper.component.shortcut
+
+//바로가기 제목 박스
+
+//바로가기 작은 박스
+
+//바로가기 큰 박스
