+++ app/src/main/java/com/ganaljigi/kubf/ui/helper/component/topappbar/TopAppBarComponent.kt
@@ -1,15 +1,11 @@
-package com.ganaljigi.kubf.ui.helper.component
+package com.ganaljigi.kubf.ui.helper.component.topappbar
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.KeyboardArrowLeft
-import androidx.compose.material.icons.filled.KeyboardArrowRight
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -25,41 +21,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
-//Helper화면의 NoticeBox
-@Composable
-fun NoticeBox(
-    onNavigateClick: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp)
-                .height(52.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = "공지사항",
-                fontWeight = FontWeight.SemiBold
-            )
-
-            IconButton(onClick = onNavigateClick) {
-                Icon(
-                    imageVector = Icons.Default.KeyboardArrowRight,
-                    contentDescription = "공지사항 페이지로 이동"
-                )
-            }
-        }
-    }
-}
-
-//Notice화면의 공지사항 TopAppBar
+//장애학생지원센터 TopAppBar
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NoticeTopAppBar(
+fun HelperTopAppBar(
     onBackClick: () -> Unit
 ) {
     TopAppBar(
@@ -68,7 +33,7 @@ fun NoticeTopAppBar(
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 Text(
-                    text = "공지사항",
+                    text = "장애학생지원센터",
                     fontSize = 20.sp,
                     fontWeight = FontWeight.SemiBold,
                     color = Color.Black,
@@ -90,14 +55,14 @@ fun NoticeTopAppBar(
         ),
         modifier = Modifier
             .fillMaxWidth()
-            .height(48.dp)
+            .height(64.dp)
     )
 }
 
 @Preview(showBackground = true)
 @Composable
-fun PreviewNoticeTopAppBar() {
-    NoticeTopAppBar (
+fun PreviewHelperTopAppBar() {
+    HelperTopAppBar(
         onBackClick = {}
     )
 }
